removes unnecessary files.
