+++ .travis.yml
@@ -4,6 +4,8 @@ go:
   - 1.8.x
   - 1.9.x
   - 1.10.x
+  - 1.11.x
+  - 1.12.x
   - tip
 
 install:
